% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calculate_feature_genomic_coverage}
\alias{calculate_feature_genomic_coverage}
\title{Calculate the number of features binned by a given size}
\usage{
calculate_feature_genomic_coverage(features, genome, binsize = 1e+05)
}
\arguments{
\item{features}{a [tibble][tibble::tibble-package] containing the columns chrom, start and end}

\item{genome}{a [tibble][tibble::tibble-package] containing the columns name and length, describing each chromosome}

\item{binsize}{the size of the bins to calculate to}
}
\value{
a [tibble][tibble::tibble-package] containing the columns Chr, Start, End, describing the bin,
  Value and ScaledValue, the total bases in that bin covered by one or more features.
}
\description{
Calculate the number of features binned by a given size
}
