% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrosynteny.R
\name{compute_cvalues}
\alias{compute_cvalues}
\title{compute chromosomal c-values}
\usage{
compute_cvalues(
  ms.x,
  ms.y,
  adjust = c("max", "minmax", "ci", "ci_mm", "none"),
  alg_genes = NULL
)
}
\arguments{
\item{ms.x}{a msynt [tibble][tibble::tibble-package]. Genes from this}

\item{ms.y}{a msynt [tibble][tibble::tibble-package]}

\item{adjust}{one of
max - divide all c-values by the maximum c-value
minmax - divide c-values by the lower maximum c-value assigned to the
         chromosome of either species. This is useful if the links
         where a normal c-value would penalize chromosomal fusions
ci - divide by the conservation index, as defined by the number of
     genes placed in ancestral linkage groups divided by the total
     number of orthologous genes placed on the genome.
ci_mm - adjust the minmax c-values by the conservation}
}
\description{
Construct a matrix of c-values, which is the number of genes that are
conserved in each pair of chromosomes, as compared to the total number
of genes in each of the chromosomes. Optionally, adjust c-values to
correct for pairs of genomes with poorly conserved chromosomes.
}
