% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_inference.R
\name{compute_gene_support}
\alias{compute_gene_support}
\title{Compute gene support for a list of clusterings}
\usage{
compute_gene_support(pool, k, method = "GV1", nruns = 5, verbose = F)
}
\arguments{
\item{pool}{A list of clusterings}

\item{k}{The number of clusters}

\item{method}{The method to use for computing gene support}

\item{nruns}{The number of runs to use for computing gene support}

\item{verbose}{Whether to print progress}
}
\value{
A tibble with the gene support values
}
\description{
Compute gene support for a list of clusterings
}
