% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrosynteny.R
\name{create_msynt}
\alias{create_msynt}
\title{create a msynt [tibble][tibble::tibble-package] for tracking macrosynteny}
\usage{
create_msynt(
  gene_positions,
  orthogroups,
  algs = NULL,
  max_group_size = NULL,
  include_chrom = NULL,
  min_scaffold_size = NULL,
  max_positions = 1
)
}
\arguments{
\item{gene_positions}{a [tibble][tibble::tibble-package] of gene positions with the
columns `chrom`, `Geneid` and `position` as documented above}

\item{orthogroups}{a [tibble][tibble::tibble-package] of gene orthogroups with
columns `Geneid` and `group` as documented above.}

\item{algs}{a [tibble][tibble::tibble-package] of ancestral linkage groups with the
columns `Geneid` and `alg`.}

\item{max_group_size}{filter orthogroups whose size are greater than this}

\item{include_chrom}{a character vector of genes to select the genes on}

\item{min_scaffold_size}{the minimum number of genes on a scaffold/chromosome to include the
scaffold/chromosome. Overridden by `include_chrom`.}

\item{max_positions}{restrict the number of times a gene may occur in the genome.
usually this should be 1.}
}
\description{
This function creates the main per-species [tibble][tibble::tibble-package]
needed to track ancestral linkage groups. This includes 5 columns:
\itemize{
  \item{chrom}{The chromosome on which the gene sits}
  \item{Geneid}{The name of the gene}
  \item{position}{The position of the gene on the chromosome}
  \item{group}{The orthology group of the gene. This column will need to correspond to
               the `group` column of another msynt table}
  \item{group_size}{the size of the orthology group. This is useful for filtering
                    suspiciously large groups.}
  \item{alg}{optionally an ancestral linkage group if provided}
}
}
