% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrosynteny.R
\name{detect_alg_genes.msynt}
\alias{detect_alg_genes.msynt}
\title{mark genes with the top `ngroups` macrosyntenic pairs}
\usage{
detect_alg_genes.msynt(ms.x, ms.y, ngroups, max_size = NULL)
}
\arguments{
\item{ms.x}{a msynt [tibble][tibble::tibble-package]. Genes from this}

\item{ms.y}{a msynt [tibble][tibble::tibble-package]}

\item{ngroups}{the number of groups to detect. The biggest pairs}
}
\description{
Mark the genes in `ms.x` with macrosyntenic group ID based on ancestral
linkage groups (ALG). An ALG in this case based on pairs of chromosomes
that have a high number of shared genes. Genes on the highest
`ngroups` pairs of chromosomes are marked as having an ALG, the rest
are given `NA`. The resulting [tibble][tibble::tibble-package] will have
the columns
\itemize{
  \item{Geneid} The name of the gene
  \item{alg} Ancestral linkage group
  }
}
