% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrosynteny.R
\name{gff_to_agora}
\alias{gff_to_agora}
\alias{exonerate_gff_to_agora}
\alias{psl_to_agora}
\title{convert a  [tibble][tibble::tibble-package] of various formats to agora gene list format}
\usage{
gff_to_agora(gff)

exonerate_gff_to_agora(gff)

psl_to_agora(psl)
}
\description{
These functions convert a tibble of another shape to a position-shaped [tibble][tibble::tibble-package]
to make it compatible with the AGORA pipeline
}
