% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_inference.R
\name{infer_cached_algs}
\alias{infer_cached_algs}
\title{Perform ancestral linkage group inference}
\usage{
infer_cached_algs(
  tree_text,
  critical_nodes,
  ortholog_source,
  dirty = F,
  cache_name = "algs",
  min_links = 1,
  min_scaffold_size = 50,
  gamma_opt_tries = 100,
  cluster_opt_tries = 1000,
  opt_beta = 0.01,
  gammas = seq(0.5, 1.5, 0.05),
  weight = c("n_links", "n_species", "total_depth"),
  use_gamma_prior = TRUE,
  n_iterations = 0,
  min_alg_group_size = 3
)
}
\arguments{
\item{tree_text}{A newick tree string}

\item{critical_nodes}{A vector of node names that must be present in the tree}

\item{ortholog_source}{Output table from OMA standalone}

\item{dirty}{If TRUE, the cache will be ignored and the inference will be run
from scratch}

\item{cache_name}{Name of the cache file to use}

\item{min_links}{Minimum number of orthologs required to infer a link}

\item{min_scaffold_size}{Minimum number of genes included in a scaffold for use in inference}

\item{gamma_opt_tries}{Number of random gamma values to try when optimizing gamma}

\item{cluster_opt_tries}{Number of random clusterings to try when optimizing clustering}

\item{opt_beta}{Beta value to use when optimizing gamma (see Leiden algorithm)}

\item{gammas}{A vector of gamma values to try when optimizing gamma}

\item{weight}{How to calculate edge weight when computing the weighted graph.
If "n_links", the number of orthologs is used. If "n_species", the number of species is used.
If "total_depth", the sum of the total tree depth for each link to the gene is used.}

\item{use_gamma_prior}{If TRUE, a prior distribution will be calculated for gamma which
is then sampled. The prior distribution is determined based on the modularity of the graphs
that are inferred when using gamma values.}

\item{n_iterations}{Number of iterations to run the inference algorithm.}

\item{min_alg_group_size}{Minimum orthologous group size to include in the
list of groups on the inferred consensus ALGs}
}
\description{
This function performs ancestral linkage group inference. The results are
cached in a file so that they can be reused.
The first step is to determine a weighted graph for each of the critical ndoes
that are indicated (e.g., "Bilaterian", "Cnidarian", "Metazoan"). The weights are
determined by the number of orthologs co-occurring on the same scaffold. Next, perform
optimized leiden clustering. The main objective of the optimization is to avoid
pre-determining the number of clusters. The optimization is performed by trying
different gamma values. The gamma value that results in the highest modularity
is used.
}
