% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrosynteny.R
\name{join_ranks}
\alias{join_ranks}
\title{Align ranks between two ranked genomes}
\usage{
join_ranks(gene_ranks.xy, gene_ranks.zy)
}
\arguments{
\item{gene_ranks.xy}{gene ranks returned from [gene_ranks()]}

\item{gene_ranks.zy}{gene ranks returned from [gene_ranks()]}
}
\value{
A list of two gene ranks [tibble][tibble::tibble-package]s with the y ranks of the y of both aligned.
}
\description{
Align ranks between two ranked genomes
}
