% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_inference.R
\name{load_cached_gene_positions}
\alias{load_cached_gene_positions}
\title{Load gene positions from annotation files}
\usage{
load_cached_gene_positions(cache_name = "pos", dirty = F)
}
\arguments{
\item{cache_name}{Name of cache file}

\item{dirty}{If TRUE, recompute the cache}
}
\value{
A list of data frames with columns chrom, Geneid, position
}
\description{
Load gene positions from annotation files
}
