% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrosynteny.R
\name{macrosynteny_cluster}
\alias{macrosynteny_cluster}
\title{Default clustering method for macrosynteny analysis}
\usage{
macrosynteny_cluster(m, priority, ...)
}
\arguments{
\item{m}{the matrix of counts common to scaffolds or chromosomes}

\item{priority}{the priority of each oto be reordered when all other things are equal. passed to reorder.dendrogram()}
}
\description{
Default clustering method for macrosynteny analysis
}
