% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrosynteny.R
\name{macrosynteny_matrix}
\alias{macrosynteny_matrix}
\title{create a matrix containing the number of genes shared in all pairs of chromosomes}
\usage{
macrosynteny_matrix(ms.x, ms.y, metric = "group")
}
\arguments{
\item{ms.x}{a msynt [tibble][tibble::tibble-package]}

\item{ms.y}{a msynt [tibble][tibble::tibble-package]}

\item{metric}{a column in the msynt (usually either `group` or `alg`) by which to make the matrix}
}
\description{
create a matrix containing the number of genes shared in all pairs of chromosomes
}
