% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrosynteny.R
\name{make_chrom_names}
\alias{make_chrom_names}
\title{Change the chromosome names to the the standard format: <SP>.<CHROM>
where <SP> is the two letter canonical species name (e.g. hs for human)
and the chromosome number <CHROM>. Chromosome numbers are in descending order of
length.}
\usage{
make_chrom_names(pos, prefix, fasta = NULL)
}
\arguments{
\item{pos}{a position-formatted [tibble][tibble::tibble-package]}

\item{prefix}{the two-letter species prefix}

\item{fasta}{optionally a fasta file to determine actual chromosome length.}
}
\description{
Change the chromosome names to the the standard format: <SP>.<CHROM>
where <SP> is the two letter canonical species name (e.g. hs for human)
and the chromosome number <CHROM>. Chromosome numbers are in descending order of
length.
}
