% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_inference.R
\name{optimized_leiden_clustering}
\alias{optimized_leiden_clustering}
\title{Optimized clustering algorithm}
\usage{
optimized_leiden_clustering(
  g,
  n_iterations = 0,
  gammas = seq(0.5, 1.5, 0.05),
  gamma_opt_tries = 100,
  cluster_opt_tries = 1000,
  use_gamma_prior = TRUE,
  opt_beta = 0.01
)
}
\arguments{
\item{g}{An igraph object}

\item{n_iterations}{Number of iterations to run the Leiden algorithm}

\item{gammas}{A vector of gamma values to try}

\item{gamma_opt_tries}{Number of times to try each gamma value}

\item{cluster_opt_tries}{Number of times to try each gamma value}

\item{use_gamma_prior}{If TRUE, use the gamma prior to sample gamma values}

\item{opt_beta}{The beta value to use for the Leiden algorithm}
}
\value{
A list with the following elements:
\itemize{
 \item modularity_data: A data frame with the modularity values for each gamma value
 \item cluster_sizes: A vector of cluster sizes for each clustering
 \item cluster_modularity: A vector of modularity values for each clustering
 \item gamma_priors: A data frame with the gamma priors
 \item gamma_fit: The gamma prior fit
 \item gamma_samples: The gamma values sampled
 \item gene_support: The gene support values
 \item gene_support_se: The gene support standard errors
 \item optimized_gamma: The optimized gamma value
 \item optimized_k: The optimized number of clusters
 \item optimized_clusters_pool: The optimized cluster pool
}
}
\description{
This function performs optimized clustering using the Leiden algorithm.
Instead of requiring a resolution (gamma) value, the algorithm tries
several gamma values and chooses the one that results in the highest
modularity. Optionally, a distribution of gamma values can be inferred
for the final `cluster_opt_tries` number of clusterings. This distribution
can then be used to sample gamma values.
}
