% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrosynteny.R
\name{plot_macrosynteny_dots}
\alias{plot_macrosynteny_dots}
\title{Make a macrosynteny dotplot}
\usage{
plot_macrosynteny_dots(
  gene_ranks,
  alg_genes = NULL,
  alg_genes_only = F,
  show_chrom_names = F,
  xtitle = "",
  ytitle = "",
  point_size = 0.2,
  shape = 16,
  linetype = "solid",
  linesize = 0.2,
  linecolor = "darkgrey",
  annotate_algs = F,
  show_legend = F,
  show_na = T,
  cols = NULL,
  nonorthologous_color = "#777777FF"
)
}
\arguments{
\item{gene_ranks}{the gene ranks [tibble][tibble::tibble-package] from [gene_ranks()]}

\item{alg_genes}{optional tibble of alg genes to mark linkage groups}

\item{alg_genes_only}{only plot genes with an alg group}

\item{show_chrom_names}{show the chromosome names on the axes. May be T, F or a vector
of 2 logicals incidating whether to show the x and y chrom names, respectively.}

\item{xtitle}{a title for the x axis (usually the species)}

\item{ytitle}{a title for the y axis}

\item{point_size}{the size of the points as passed to [ggplot2::geom_point]}

\item{shape}{the shape of points as passed to [ggplot2::geom_point]}

\item{annotate_algs}{show names of the ancestral linkage groups by color}

\item{nonorthologous_color}{genes which do not fit into an ancestral linkage group}
}
\description{
Make a macrosynteny dotplot
}
