% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrosynteny.R
\name{plot_macrosynteny_heatmap}
\alias{plot_macrosynteny_heatmap}
\title{plot heatmap of macrosynteny}
\usage{
plot_macrosynteny_heatmap(
  ms.x,
  ms.y,
  metric = "group",
  clust_method = macrosynteny_cluster,
  scale = F,
  ...
)
}
\arguments{
\item{ms.x}{a msynt [tibble][tibble::tibble-package]}

\item{ms.y}{a msynt [tibble][tibble::tibble-package]}

\item{metric}{a column in msynt (usually either `group` or `alg`) by which the pairs of chromosomes are measured}

\item{scale}{whether to scale the initial matrix. This works better for comparing to scaffolded genomes.}

\item{...}{arguments passed to [ComplexHeatmap::Heatmap()]}
}
\description{
Plots a [ComplexHeatmap::Heatmap] corresponding to the number of genes
shared by each pair of chromosomes
}
