% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrosynteny.R
\name{psl_to_position}
\alias{psl_to_position}
\alias{msynt_to_position}
\alias{exonerate_gff_to_position}
\title{convert a  [tibble][tibble::tibble-package] of various formats to position format}
\usage{
psl_to_position(psl)

msynt_to_position(msynt)

exonerate_gff_to_position(gff)
}
\description{
These functions convert a tibble of another shape to a position-shaped [tibble][tibble::tibble-package]
to make it compatible with [create_msynt()]
}
