% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrosynteny.R
\name{rank_genes}
\alias{rank_genes}
\title{Rank genes across the genome based on macrosynnteny and chromosome position}
\usage{
rank_genes(
  ms.x,
  ms.y,
  metric = "group",
  order.x = F,
  order.y = F,
  clust_method = macrosynteny_cluster
)
}
\arguments{
\item{ms.x}{a msynt [tibble][tibble::tibble-package].}

\item{ms.y}{a msynt [tibble][tibble::tibble-package].}

\item{metric}{a column in msynt (usually either `group` or `alg`) by which the genes are ranked}

\item{order.x}{pass a customized ordering of genes for the x axis}

\item{order.y}{pass a customized ordering of genes for the y axis}

\item{clust_method}{a function accepting the matrix and an additional priority parameter (which may be ignored) for dendrogram-based methods}
}
\description{
Assign a position in the genome by discretizing chromosomes to one gene per
position from 1 to the number of genes in the genome. The genes will be ordered
based on both similarity in number of shared genes per chromosome pair, then by
position in the chromosome
}
