% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrosynteny.R
\name{seriate_macrosynteny_cluster_method}
\alias{seriate_macrosynteny_cluster_method}
\title{Generate a seriation based method}
\usage{
seriate_macrosynteny_cluster_method(method = "TSP", dist = T, ...)
}
\arguments{
\item{method}{the method to be used. see the seriation package for details}

\item{dist}{whether to convert the matrix to a distance before passing to seriat()}
}
\description{
This generates a method like macrosynteny_cluster but based on a chosen method in from the seriation package.
}
