% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{which_next}
\alias{which_next}
\title{Return the next index of `x` which satisfies a predicate `p()`}
\usage{
which_next(x, p, cur = 1)
}
\arguments{
\item{x}{a vector}

\item{p}{a predicate for which the next index should be true}

\item{cur}{the starting index}
}
\value{
the first index `i` after `cur` for which `p(x[i])` is true
}
\description{
The advantage of using this function over `which()` is that it does not
calculate for all values in the vector `x`, offering a very big speedup.
}
\examples{
  which_next(seq(10), function(x) x > 5)
  which_next(seq(10), function(x) x > 0)
}
