% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrosynteny.R
\name{write_agora}
\alias{write_agora}
\title{Write an AGORA [tibble][tibble::tibble-package] in to a path}
\usage{
write_agora(agora, path)
}
\arguments{
\item{agora}{a [tibble][tibble::tibble-package] containing the columns chrom, Geneid, position, end and strand}

\item{path}{the output path of the file}
}
\description{
Write an AGORA [tibble][tibble::tibble-package] in to a path
}
